+++ enemyMovement.js
@@ -1,0 +1,44 @@
+const stdin = process.stdin;
+stdin.setRawMode(true);
+stdin.resume();
+stdin.setEncoding('utf8');
+
+const iceMovement = require('./iceAlteration')
+const matrixFunctions = require('./matrixFunctions');
+const objects = require('./objects');
+const fs = require('fs');
+
+const enemy = objects.enemy;
+const turnEnemy = (enemy, direction) => {
+  enemy.direction = direction;
+};
+
+const randomMove = () => {
+  return Math.floor(Math.random() * 4);
+};
+
+const moveEnemy = (enemy, direction, matrix) => {
+  if (randomMove() === 0 && enemy.xCoord > 0 && matrix[enemy.xCoord - 1][enemy.yCoord] !== '1') { // up
+    enemy.xCoord--;
+    matrix[enemy.xCoord + 1][enemy.yCoord] = ' ';
+  } else if (randomMove() === 1 && enemy.xCoord < matrix.length - 1 && matrix[enemy.xCoord + 1][enemy.yCoord] !== '1') { // down
+    enemy.xCoord++;
+    matrix[enemy.xCoord - 1][enemy.yCoord] = ' ';
+  } else if (randomMove() === 2 && enemy.yCoord > 0 && matrix[enemy.xCoord][enemy.yCoord - 1] !== '1') { // left
+    enemy.yCoord--;
+    matrix[enemy.xCoord][enemy.yCoord + 1] = ' ';
+  } else if (randomMove() === 3 && enemy.yCoord < matrix[0].length - 1 && matrix[enemy.xCoord][enemy.yCoord + 1] !== '1') { // right
+    enemy.yCoord++;
+    matrix[enemy.xCoord][enemy.yCoord - 1] = ' ';
+  }
+};
+
+
+//collision (implement in moveEnemy pls.):
+
+
+const collision = (obj1, obj2) => {
+  if (obj1.type === 'slidingBlock' && obj2.type === 'enemy') {
+    obj2 = { type: 'blood', symbol: objects.blood.symbol };
+  }
+}
